--- java-apm-agent/java-agent-extension/src/main/java/org/traffichunter/javaagent/extension/InstrumentationHelper.java
@@ -1,72 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright (c) 2024 traffic-hunter.org
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.traffichunter.javaagent.extension;
-
-import java.util.List;
-import java.util.logging.Logger;
-import net.bytebuddy.agent.builder.AgentBuilder;
-import net.bytebuddy.agent.builder.AgentBuilder.Identified.Extendable;
-import org.traffichunter.javaagent.extension.bytebuddy.AdjustTransformer;
-
-/**
- * @author yungwang-o
- * @version 1.1.0
- */
-public class InstrumentationHelper {
-
-    private static final Logger log = Logger.getLogger(InstrumentationHelper.class.getName());
-
-    private final List<AbstractPluginInstrumentation> pluginInstrumentation;
-
-    public InstrumentationHelper(final List<AbstractPluginInstrumentation> pluginInstrumentation) {
-        this.pluginInstrumentation = pluginInstrumentation;
-    }
-
-    public AgentBuilder instrument(AgentBuilder originalAgentBuilder) {
-
-        if(isEnabled()) {
-            log.warning("Instrumenting is empty!");
-            return originalAgentBuilder;
-        }
-
-        for(final AbstractPluginInstrumentation plugin : pluginInstrumentation) {
-
-            Extendable transform = originalAgentBuilder
-                    .type(plugin.typeMatcher())
-                    .transform(new AdjustTransformer());
-
-            Transformer transformer = new Transformer(transform);
-
-            plugin.transform(transformer);
-
-            originalAgentBuilder = transformer.agentBuilder();
-        }
-
-        return originalAgentBuilder;
-    }
-
-    private boolean isEnabled() {
-        return pluginInstrumentation == null || pluginInstrumentation.isEmpty();
-    }
-}
