+++ java-apm-agent/java-agent-extension/src/main/java/org/traffichunter/javaagent/extension/TraceInfo.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.javaagent.extension.otel;
+package org.traffichunter.javaagent.extension;
 
 import io.opentelemetry.api.trace.StatusCode;
 import io.opentelemetry.sdk.trace.data.SpanData;
