--- java-apm-agent/java-agent-extension/src/main/java/org/traffichunter/javaagent/extension/bootstrap/OpenTelemetryManager.java
@@ -1,60 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright (c) 2024 traffic-hunter.org
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.traffichunter.javaagent.extension.bootstrap;
-
-import io.opentelemetry.sdk.OpenTelemetrySdk;
-import io.opentelemetry.sdk.trace.IdGenerator;
-import io.opentelemetry.sdk.trace.SdkTracerProvider;
-import io.opentelemetry.sdk.trace.export.SimpleSpanProcessor;
-import io.opentelemetry.sdk.trace.export.SpanExporter;
-import io.opentelemetry.sdk.trace.samplers.Sampler;
-import org.traffichunter.javaagent.extension.otel.TrafficHunterExporter;
-
-/**
- * @author yungwang-o
- * @version 1.1.0
- */
-final class OpenTelemetryManager {
-
-    private OpenTelemetryManager() {}
-
-    static final SpanExporter spanExporter = new TrafficHunterExporter();
-
-    static OpenTelemetrySdk manageOpenTelemetrySdk() {
-
-        SdkTracerProvider provider = SdkTracerProvider.builder()
-                .addSpanProcessor(SimpleSpanProcessor.create(spanExporter))
-                .setIdGenerator(IdGenerator.random())
-                .setSampler(Sampler.alwaysOn())
-                .build();
-
-        return OpenTelemetrySdk.builder()
-                .setTracerProvider(provider)
-                .buildAndRegisterGlobal();
-    }
-
-    public static void flush() {
-        spanExporter.flush();
-    }
-}
