+++ java-apm-agent/java-agent-extension/src/main/java/org/traffichunter/javaagent/extension/property/TrafficHunterAgentProperty.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.javaagent.extension.bootstrap.property;
+package org.traffichunter.javaagent.extension.property;
 
 import java.util.concurrent.TimeUnit;
 import org.traffichunter.javaagent.retry.backoff.BackOffPolicy;
